change controlled failure endpoint to avoid netapp cert requirement
